Druid   Add 3.1.2 version of (Glyph of) Innervate   Fix level 80 base stats for druid
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@2303 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ simcraft.h
@@ -748,8 +748,8 @@ struct player_t
     int       focus_magic_feedback;
     double    fortitude;
     double    improved_divine_spirit;
-    int       innervate;
-    int       glyph_of_innervate;
+    double    innervate;
+    double    glyph_of_innervate;
     int       hysteria;
     double    mana_cost_reduction;
     double    mana_spring;
